refactor: Move ExcelFile type to own module
