+++ src/types/mod.rs
@@ -1,2 +1,4 @@
+pub(crate) mod excelfile;
 pub(crate) mod excelsheet;
+pub(crate) use excelfile::{ExcelFile, ExcelSheetIterator};
 pub(crate) use excelsheet::ExcelSheet;
